+++ src/model/Request.java
@@ -18,8 +18,8 @@ public class Request {
     AID originalSender;
     Set<Integer> allReceivers;
     // id, lifetime
-    Map<String, Integer> reservedItems;
-    int timeout;
+    Map<String, Long> reservedItems;
+    long timeout;
     boolean processed;
 
     public Request(String id, long quantity, ResourceType resourceType, Map<Long, Long> utilityFunction, AID sender) {
@@ -30,7 +30,7 @@ public class Request {
         this.sender = sender;
     }
 
-    public Request(String id, Boolean cascaded, long quantity, ResourceType resourceType, Map<Long, Long> utilityFunction, AID sender, AID originalSender, Set<Integer> allReceivers, Map<String, Integer> reservedItems) {
+    public Request(String id, Boolean cascaded, long quantity, ResourceType resourceType, Map<Long, Long> utilityFunction, AID sender, AID originalSender, Set<Integer> allReceivers, Map<String, Long> reservedItems) {
         this.id = id;
         this.cascaded = cascaded;
         this.quantity = quantity;
@@ -42,7 +42,7 @@ public class Request {
         this.reservedItems = reservedItems;
     }
 
-    public Request(String id, Boolean cascaded, long quantity, ResourceType resourceType, Map<Long, Long> utilityFunction, AID sender, AID originalSender, Set<Integer> allReceivers, Map<String, Integer> reservedItems, int timeout) {
+    public Request(String id, Boolean cascaded, long quantity, ResourceType resourceType, Map<Long, Long> utilityFunction, AID sender, AID originalSender, Set<Integer> allReceivers, Map<String, Long> reservedItems, long timeout) {
         this.id = id;
         this.cascaded = cascaded;
         this.quantity = quantity;
