+++ src/model/Task.java
@@ -6,11 +6,11 @@ public class Task {
 
     String id;
     long utility;
-    int deadline;
+    long deadline;
     Map<ResourceType, Long> requiredResources;
 //    HashMap<ResourceType, HashMap<Integer, TimeInterval>> requiredResources;
 
-    public Task(String id, long utility, int deadline, Map<ResourceType, Long> requiredResources) {
+    public Task(String id, long utility, long deadline, Map<ResourceType, Long> requiredResources) {
         this.id = id;
         this.utility = utility;
         this.deadline = deadline;
